# Projekat 1/gateway service/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "gateway service",
    "version": "1.0"
  },
  "paths": {
    "/Gateway": {
      "get": {
        "tags": [
          "Gateway"
        ],
        "operationId": "GetLapTimeWeatherForecast",
        "parameters": [
          {
            "name": "driverId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "raceId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "lap",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Gateway"
        ],
        "operationId": "AddLapTime",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LapTime"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LapTime"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LapTime"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Gateway"
        ],
        "operationId": "UpdateLapTime",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LapTime"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LapTime"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LapTime"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Gateway"
        ],
        "operationId": "DeleteLapTime",
        "parameters": [
          {
            "name": "driverId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "raceId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "lap",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LapTime": {
        "type": "object",
        "properties": {
          "raceId": {
            "type": "integer",
            "format": "int32"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "round": {
            "type": "integer",
            "format": "int32"
          },
          "circuitId": {
            "type": "integer",
            "format": "int32"
          },
          "race_name": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "nullable": true
          },
          "race_time": {
            "type": "number",
            "format": "double"
          },
          "ciruit_name": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lng": {
            "type": "number",
            "format": "double"
          },
          "driverId": {
            "type": "integer",
            "format": "int32"
          },
          "forename": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          },
          "nationality": {
            "type": "string",
            "nullable": true
          },
          "lap": {
            "type": "integer",
            "format": "int32"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "lap_time": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      }
    }
  }
}